feat: support target param
