+++ packages/plugin-ice-ssr/src/index.ts
@@ -1,20 +1,35 @@
 import * as path from 'path';
+import * as fse from 'fs-extra';
+import { minify } from 'html-minifier';
 import LoadablePlugin from '@loadable/webpack-plugin';
 import getWebpackConfig from '@builder/webpack-config';
-import setSSGWebpackConfig from './setSSGWebpackConfig';
-import setSSRWebpackConfig from './setSSRWebpackConfig';
+import { formatPath } from '@builder/app-helpers';
+import SSGWebpackPlugin from './SSGWebpackPlugin';
 
 const plugin = async (api): Promise<void> => {
-  const { context, onGetWebpackConfig, modifyUserConfig } = api;
+  const { context, registerTask, getValue, onGetWebpackConfig, onHook, log, applyMethod, modifyUserConfig } = api;
   const { rootDir, command, webpack, commandArgs, userConfig } = context;
-  const { outputDir, ssr } = userConfig;
-
+  const { outputDir, target = 'server',  publicPath = '/', devPublicPath = '/' } = userConfig;
+  const PROJECT_TYPE = getValue('PROJECT_TYPE');
+  const TEMP_PATH = getValue('TEMP_PATH');
   // Note: Compatible plugins to modify configuration
   const buildDir = path.join(rootDir, outputDir);
   const serverDir = path.join(buildDir, 'server');
+  const serverFilePath = path.join(serverDir, 'server.js');
 
-  // const serverFilename = 'index.js';
-  // const serverFilePath = path.join(serverDir, serverFilename);
+  // render server entry
+  const templatePath = path.join(__dirname, '../src/server.ts.ejs');
+  const ssrEntry = path.join(TEMP_PATH, 'server.ts');
+  const routesFileExists = fse.existsSync(path.join(rootDir, 'src', `routes.${PROJECT_TYPE}`));
+  applyMethod(
+    'addRenderFile',
+    templatePath,
+    ssrEntry,
+    {
+      outputDir,
+      routesPath: routesFileExists ? '@' : '.',
+      publicPath: command === 'build' ? publicPath : devPublicPath
+    });
 
   const mode = command === 'start' ? 'development' : 'production';
 
@@ -50,23 +65,173 @@ const plugin = async (api): Promise<void> => {
       '@loadable/babel-plugin',
     ]
   );
+  registerTask('ssr', webpackConfig);
+  onGetWebpackConfig('ssr', (config) => {
+    config.entryPoints.clear();
 
-  if (ssr) {
-    // ssr webpack node config
-    setSSRWebpackConfig({ api, mode, serverDir, buildDir });
-  } else {
-    // ssg webpack node config
-    setSSGWebpackConfig({ api, mode, serverDir, buildDir });
-  }
+    config.entry('server').add(ssrEntry);
 
-  // onHook(`after.${command}.compile`, () => {
-  //   const htmlFilePath = path.join(buildDir, 'index.html');
-  //   const bundle = fse.readFileSync(serverFilePath, 'utf-8');
-  //   const html = fse.readFileSync(htmlFilePath, 'utf-8');
-  //   const minifiedHtml = minify(html, { collapseWhitespace: true, quoteCharacter: '\'' });
-  //   const newBundle = bundle.replace(/__ICE_SERVER_HTML_TEMPLATE__/, minifiedHtml);
-  //   fse.writeFileSync(serverFilePath, newBundle, 'utf-8');
-  // });
+    config.target('node');
+
+    config.name('ssr');
+
+    config.module
+      .rule('polyfill')
+      .include.add(ssrEntry);
+
+    config
+      .plugin('DefinePlugin')
+      .use(webpack.DefinePlugin)
+      .tap(([args]) => [{ ...args, 'process.env.__IS_SERVER__': true }]);
+
+    config.plugins.delete('MiniCssExtractPlugin');
+    ['scss', 'scss-module', 'css', 'css-module', 'less', 'less-module'].forEach((rule) => {
+      if (config.module.rules.get(rule)) {
+        config.module.rule(rule).uses.delete('MiniCssExtractPlugin.loader');
+        config.module
+          .rule(rule)
+          .use('css-loader')
+          .tap((options) => ({
+            ...options,
+            modules: {
+              ...(options.modules || {}),
+              exportOnlyLocals: true,
+            },
+          }));
+      }
+    });
+
+    config.output
+      .path(serverDir)
+      .filename('[name].js')
+      .publicPath('/')
+      .libraryTarget('commonjs2');
+
+    // in case of app with client and server code, webpack-node-externals is helpful to reduce server bundle size
+    // while by bundle all dependencies, developers do not need to concern about the dependencies of server-side
+    // TODO: support options to enable nodeExternals
+    // empty externals added by config external
+    config.externals([]);
+
+    // remove process fallback when target is node
+    config.plugins.delete('ProvidePlugin');
+
+    async function serverRender(res, req) {
+      const htmlTemplate = fse.readFileSync(path.join(buildDir, 'index.html'), 'utf8');
+      log.info('[SSR]', 'start server render');
+      delete require.cache[serverFilePath];
+      // eslint-disable-next-line
+      const serverRender = require(serverFilePath);
+      const ctx = { res, req };
+      const { error, html, redirectUrl } = await serverRender.default(ctx, { htmlTemplate });
+
+      if (redirectUrl) {
+        log.info('[SSR]', `Redirect to the new path ${redirectUrl}`);
+        res.redirect(302, redirectUrl);
+      } else {
+        if (error) {
+          log.error('[SSR] Server side rendering error, downgraded to client side rendering');
+          log.error(error);
+        }
+        log.info('[SSR] SSR success', 'output html content');
+        log.verbose('[SSR] ssr html content', html);
+        res.send(html);
+      }
+    }
+
+    if (command === 'start') {
+      const originalDevMiddleware = config.devServer.get('devMiddleware');
+      config.devServer.set('devMiddleware', {
+        ...originalDevMiddleware,
+        writeToDisk: (filePath: string) => {
+          const formattedFilePath = formatPath(filePath);
+          return /(server\/.*|loadable-stats.json|index.html)$/.test(formattedFilePath);
+        },
+      });
+
+      let serverReady = false;
+      let httpResponseQueue = [];
+      const originalDevServeBefore = config.devServer.get('onBeforeSetupMiddleware');
+      config.devServer.set('onBeforeSetupMiddleware', (server) => {
+        const { app } = server;
+        if (typeof originalDevServeBefore === 'function') {
+          originalDevServeBefore(server);
+        }
+        let compilerDoneCount = 0;
+        server.compiler.compilers.forEach((compiler) => {
+          compiler.hooks.done.tap('ssrServer', () => {
+            compilerDoneCount++;
+            // wait until all compiler is done
+            if (compilerDoneCount === server.compiler.compilers.length) {
+              serverReady = true;
+              httpResponseQueue.forEach(([req, res]) => {
+                serverRender(res, req);
+              });
+              // empty httpResponseQueue
+              httpResponseQueue = [];
+            }
+          });
+        });
+
+        const pattern = /^\/?((?!\.(js|css|map|json|png|jpg|jpeg|gif|svg|eot|woff2|ttf|ico)).)*$/;
+        app.get(pattern, async (req, res) => {
+          if (serverReady) {
+            serverRender(res, req);
+          } else {
+            httpResponseQueue.push([req, res]);
+          }
+        });
+      });
+    }
+
+    if (command === 'build' && target === 'static') {
+      // enable ssg
+      const ssgServerTemplatePath = path.join(__dirname, './ssgServer.ts.ejs');
+      const ssgServerEntryPath = path.join(TEMP_PATH, './ssgServer.ts');
+      const ssgServerBundlePath = path.join(serverDir, 'ssgServer.js');
+      applyMethod(
+        'addRenderFile',
+        ssgServerTemplatePath,
+        ssgServerEntryPath,
+        {
+          outputDir,
+          routesPath: routesFileExists ? '@' : '.',
+          publicPath: command === 'build' ? publicPath : devPublicPath
+        }
+      );
+
+      const renderTemplatePath = path.join(__dirname, './ssgRender.ts.ejs');
+      const renderEntryPath = path.join(TEMP_PATH, 'ssgRender.ts');
+      applyMethod(
+        'addRenderFile',
+        renderTemplatePath,
+        renderEntryPath
+      );
+
+      config.entry('ssgServer').add(ssgServerEntryPath);
+      config.entry('ssgRender').add(renderEntryPath);
+
+      config
+        .plugin('SSGWebpackPlugin')
+        .after('HtmlWebpackPlugin')
+        .use(SSGWebpackPlugin, [
+          {
+            serverFilePath: ssgServerBundlePath,
+            buildDir,
+            command
+          }
+        ]);
+    }
+  });
+
+  onHook(`after.${command}.compile`, () => {
+    const htmlFilePath = path.join(buildDir, 'index.html');
+    const bundle = fse.readFileSync(serverFilePath, 'utf-8');
+    const html = fse.readFileSync(htmlFilePath, 'utf-8');
+    const minifiedHtml = minify(html, { collapseWhitespace: true, quoteCharacter: '\'' });
+    const newBundle = bundle.replace(/__ICE_SERVER_HTML_TEMPLATE__/, minifiedHtml);
+    fse.writeFileSync(serverFilePath, newBundle, 'utf-8');
+  });
 };
 
 export default plugin;
